refactor: remove web org metadata leftovers
