+++ turbo/apps/web/src/__tests__/test-constants/org.ts
@@ -1,0 +1,1 @@
+export const ORG_SENTINEL_USER_ID = "__org__";
