--- turbo/apps/web/src/lib/zero/org/org-sentinel.ts
@@ -1,2 +1,0 @@
-/** Sentinel userId for org-level resources (model providers, secrets, variables) */
-export const ORG_SENTINEL_USER_ID = "__org__";
